typed-protocols: Data.Type.Queue
* move type level Queue and its singletons to 'Data.Type.Queue'.
* add Sing and SingI instances
