+++ typed-protocols/typed-protocols.cabal
@@ -24,6 +24,7 @@ library
                    , Network.TypedProtocol.Codec
                    , Network.TypedProtocol.Driver
                    , Network.TypedProtocol.Proofs
+                   , Data.Type.Queue
 
   other-extensions:  GADTs
                    , RankNTypes
